--- src/app/(routes)/chat/page.tsx
@@ -1,6 +1,0 @@
-'use cleint';
-import ChatComponent from '@/components/chat/chat';
-
-export default function Chat() {
-  return <ChatComponent></ChatComponent>;
-}
