+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 import './globals.css';
 import Provider from '../providers/_provider';
-import ChatRoomList from '@/components/chat/chatRoomList';
+import ChatRoomList from '@/components/chat/chat';
 // import Analytics from '@/components/analytics/page';
 import GNB from '@/components/layout/gnb/template';
