+++ src/components/chat/types/chat.type.ts
@@ -1,3 +1,5 @@
+import { PropsWithChildren } from 'react';
+import { UserType } from '@/components/authPage/common.types';
 export interface MessageResponse {
   content: string;
   createdAt: Date;
@@ -7,4 +9,32 @@ export interface MessageResponse {
 export interface RowProps {
   index: number;
   style?: React.CSSProperties;
+}
+
+export interface ChatList extends PropsWithChildren {
+  isOpen?: boolean;
+}
+export interface ChatRoomType {
+  conversationPartner: {
+    joinedAt: Date;
+    userId: string;
+    role: string;
+    user: {
+      name: string;
+      userType: UserType;
+    };
+  }[];
+  id: string;
+  roomCreatedAt: Date;
+  lastMessage: LastMessageType;
+}
+
+export interface LastMessageType {
+  id: string;
+  content: string;
+  updatedAt: Date;
+  user: {
+    name: string;
+    userType: UserType;
+  };
 }
